refactor(tui): split the retained backend by concern (3,557 → 1,001 lines)
`retained.rs` was the second-largest file in the crate, and a third of it was an inline test
module. What is left is the part that has to be one thing: the state, the command enum, and the
loop that applies one to the other.

  retained/tests.rs    935  was inline; unchanged apart from dedenting one level
  retained/paint.rs    639  one frame: transcript · highlight · working line · footer + input row
  retained/api.rs      333  what other threads may do: start/stop/suspend + the command senders
  retained/widgets.rs  261  overlay · assistant rows · tool row · plan box · diff box · verify line
  retained/geom.rs     228  the geometry slots, mouse hit-testing, selection extraction
  retained/ansi.rs     214  display-column measuring/clipping, and ANSI stripping/parsing

The submodules use `use super::*` rather than naming imports, and that is deliberate here where it
would not have been for `main.rs`: these are not separate subsystems that happened to share a file,
they are one renderer sharing one `AppState` and one ratatui prelude. `ansi.rs` is the exception —
it is genuinely self-contained, so it imports the two ratatui types it needs and nothing else.

Visibility was measured, not guessed. Moving an item out of the module forces its visibility to be
stated, and a first pass published all 86; measuring what actually reaches across a boundary put 11
back to private and 24 down to `pub(super)`, leaving 51 genuinely `pub(crate)`. Four more looked
external only because their names (`send`, `draw`, `draw_transcript`, `working_line`) collide with
functions elsewhere — no caller writes `retained::<any of them>`, so they are `pub(super)` too. The
`private_interfaces` lint is what caught the over-exposure: a `pub(crate)` function taking
`&mut AppState` leaks a private type, which is the compiler saying the boundary was drawn wrong.

Also corrected the module doc, which promised a rollback path that no longer exists: it said the
legacy sticky renderer in the parent module "remains the rollback path", but that renderer was
deleted — retained is the only chat surface aizen has.

Verified: 1635 passed / 0 failed / 2 ignored, so all 935 moved test lines still run; fmt clean;
clippy 125 warnings, the baseline, with the single one inside retained/ being a pre-existing lint
that moved with its code.
